+++ app/lib/bk/compat/models/plugin.rb
