+++ app/lib/bk/compat/parsers/circleci/docker.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../../pipeline/plugin'
+require_relative '../../models/plugin'
 require_relative '../../pipeline/step'
 
 module BK
